+++ src/CompilerKit/src/Assemblers/Assembler+32x0.cc
@@ -34,6 +34,6 @@
 /////////////////////////////////////////////////////////////////////////////////////////
 
 NECTI_MODULE(NEAssemblerMain32000) {
-  CompilerKit::install_signal(SIGSEGV, Detail::drvi_crash_handler);
+  CompilerKit::install_signal(SIGSEGV, CompilerKit::Detail::drvi_crash_handler);
   return EXIT_SUCCESS;
 }
