+++ src/CompilerKit/src/Assemblers/Assembler+PowerPC.cc
@@ -73,7 +73,7 @@ static bool asm_read_attributes(std::string line);
 /////////////////////////////////////////////////////////////////////////////////////////
 
 NECTI_MODULE(AssemblerMainPower64) {
-  CompilerKit::install_signal(SIGSEGV, Detail::drvi_crash_handler);
+  CompilerKit::install_signal(SIGSEGV, CompilerKit::Detail::drvi_crash_handler);
 
   for (size_t i = 1; i < argc; ++i) {
     if (argv[i][0] == '-') {
@@ -149,7 +149,7 @@ NECTI_MODULE(AssemblerMainPower64) {
 
     while (std::getline(file_ptr, line)) {
       if (auto ln = asm64.CheckLine(line, argv[i]); !ln.empty()) {
-        Detail::print_error(ln, argv[i]);
+        CompilerKit::Detail::print_error(ln, argv[i]);
         continue;
       }
 
@@ -159,7 +159,7 @@ NECTI_MODULE(AssemblerMainPower64) {
       } catch (const std::exception& e) {
         if (kVerbose) {
           std::string what = e.what();
-          Detail::print_warning("exit because of: " + what, "CompilerKit");
+          CompilerKit::Detail::print_warning("exit because of: " + what, "CompilerKit");
         }
 
         std::filesystem::remove(object_output);
@@ -274,14 +274,14 @@ static bool asm_read_attributes(std::string line) {
   // that we need this symbol.
   if (CompilerKit::find_word(line, "extern_segment")) {
     if (kOutputAsBinary) {
-      Detail::print_error("Invalid extern_segment directive in flat binary mode.", "CompilerKit");
+      CompilerKit::Detail::print_error("Invalid extern_segment directive in flat binary mode.", "CompilerKit");
       throw std::runtime_error("invalid_extern_segment_bin");
     }
 
     auto name = line.substr(line.find("extern_segment") + strlen("extern_segment") + 1);
 
     if (name.size() == 0) {
-      Detail::print_error("Invalid extern_segment", "CompilerKit");
+      CompilerKit::Detail::print_error("Invalid extern_segment", "CompilerKit");
       throw std::runtime_error("invalid_extern_segment");
     }
 
@@ -333,7 +333,7 @@ static bool asm_read_attributes(std::string line) {
   // .zero64
   else if (CompilerKit::find_word(line, "public_segment")) {
     if (kOutputAsBinary) {
-      Detail::print_error("Invalid public_segment directive in flat binary mode.", "CompilerKit");
+      CompilerKit::Detail::print_error("Invalid public_segment directive in flat binary mode.", "CompilerKit");
       throw std::runtime_error("invalid_public_segment_bin");
     }
 
@@ -395,7 +395,7 @@ static bool asm_read_attributes(std::string line) {
 
 // \brief algorithms and helpers.
 
-namespace Detail::algorithm {
+namespace CompilerKit::Detail::algorithm {
 // \brief authorize a brief set of characters.
 static inline bool is_not_alnum_space(char c) {
   return !(isalpha(c) || isdigit(c) || (c == ' ') || (c == '\t') || (c == ',') || (c == '(') ||
@@ -406,7 +406,7 @@ static inline bool is_not_alnum_space(char c) {
 bool is_valid_power64(std::string str) {
   return std::find_if(str.begin(), str.end(), is_not_alnum_space) == str.end();
 }
-}  // namespace Detail::algorithm
+}  // namespace CompilerKit::Detail::algorithm
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
@@ -426,7 +426,7 @@ std::string CompilerKit::EncoderPowerPC::CheckLine(std::string line, std::string
       line.erase(line.find(';'));
     } else {
       /// does the line contains valid input?
-      if (!Detail::algorithm::is_valid_power64(line)) {
+      if (!CompilerKit::Detail::algorithm::is_valid_power64(line)) {
         err_str = "Line contains non alphanumeric characters.\nhere -> ";
         err_str += line;
       }
@@ -435,7 +435,7 @@ std::string CompilerKit::EncoderPowerPC::CheckLine(std::string line, std::string
     return err_str;
   }
 
-  if (!Detail::algorithm::is_valid_power64(line)) {
+  if (!CompilerKit::Detail::algorithm::is_valid_power64(line)) {
     err_str = "Line contains non alphanumeric characters.\nhere -> ";
     err_str += line;
 
@@ -525,7 +525,7 @@ bool CompilerKit::EncoderPowerPC::WriteNumber(const std::size_t& pos, std::strin
     case 'x': {
       if (auto res = strtol(jump_label.substr(pos + 2).c_str(), nullptr, 16); !res) {
         if (errno != 0) {
-          Detail::print_error("invalid hex number: " + jump_label, "CompilerKit");
+          CompilerKit::Detail::print_error("invalid hex number: " + jump_label, "CompilerKit");
           throw std::runtime_error("invalid_hex");
         }
       }
@@ -546,7 +546,7 @@ bool CompilerKit::EncoderPowerPC::WriteNumber(const std::size_t& pos, std::strin
     case 'b': {
       if (auto res = strtol(jump_label.substr(pos + 2).c_str(), nullptr, 2); !res) {
         if (errno != 0) {
-          Detail::print_error("invalid binary number: " + jump_label, "CompilerKit");
+          CompilerKit::Detail::print_error("invalid binary number: " + jump_label, "CompilerKit");
           throw std::runtime_error("invalid_bin");
         }
       }
@@ -566,7 +566,7 @@ bool CompilerKit::EncoderPowerPC::WriteNumber(const std::size_t& pos, std::strin
     case 'o': {
       if (auto res = strtol(jump_label.substr(pos + 2).c_str(), nullptr, 7); !res) {
         if (errno != 0) {
-          Detail::print_error("invalid octal number: " + jump_label, "CompilerKit");
+          CompilerKit::Detail::print_error("invalid octal number: " + jump_label, "CompilerKit");
           throw std::runtime_error("invalid_octal");
         }
       }
@@ -616,7 +616,7 @@ bool CompilerKit::EncoderPowerPC::WriteNumber(const std::size_t& pos, std::strin
 
 bool CompilerKit::EncoderPowerPC::WriteLine(std::string line, std::string file) {
   if (CompilerKit::find_word(line, "public_segment")) return false;
-  if (!Detail::algorithm::is_valid_power64(line)) return false;
+  if (!CompilerKit::Detail::algorithm::is_valid_power64(line)) return false;
 
   for (auto& opcode_risc : kOpcodesPowerPC) {
     // strict check here
@@ -676,7 +676,7 @@ bool CompilerKit::EncoderPowerPC::WriteLine(std::string line, std::string file) 
               // something like r190 doesn't exist in the instruction set.
               if (isdigit(line[line_index + 3]) && isdigit(line[line_index + 2])) {
                 reg_str += line[line_index + 3];
-                Detail::print_error("invalid register index, r" + reg_str +
+                CompilerKit::Detail::print_error("invalid register index, r" + reg_str +
                                         "\nnote: The POWER accepts registers from r0 to r32.",
                                     file);
                 throw std::runtime_error("invalid_register_index");
@@ -686,7 +686,7 @@ bool CompilerKit::EncoderPowerPC::WriteLine(std::string line, std::string file) 
               std::size_t reg_index = strtol(reg_str.c_str(), nullptr, 10);
 
               if (reg_index > kAsmRegisterLimit) {
-                Detail::print_error("invalid register index, r" + reg_str, file);
+                CompilerKit::Detail::print_error("invalid register index, r" + reg_str, file);
                 throw std::runtime_error("invalid_register_index");
               }
 
@@ -707,7 +707,7 @@ bool CompilerKit::EncoderPowerPC::WriteLine(std::string line, std::string file) 
                 // check if bigger than two.
                 for (size_t i = 2; i < 4; i++) {
                   if (num.number[i] > 0) {
-                    Detail::print_warning("number overflow on li operation.", file);
+                    CompilerKit::Detail::print_warning("number overflow on li operation.", file);
                     break;
                   }
                 }
@@ -783,7 +783,7 @@ bool CompilerKit::EncoderPowerPC::WriteLine(std::string line, std::string file) 
                 ++found_some_count;
 
                 if (found_some_count > 3) {
-                  Detail::print_error("Too much registers. -> " + line, file);
+                  CompilerKit::Detail::print_error("Too much registers. -> " + line, file);
                   throw std::runtime_error("too_much_regs");
                 }
               }
@@ -792,7 +792,7 @@ bool CompilerKit::EncoderPowerPC::WriteLine(std::string line, std::string file) 
                 ++found_some_count;
 
                 if (found_some_count > 3) {
-                  Detail::print_error("Too much registers. -> " + line, file);
+                  CompilerKit::Detail::print_error("Too much registers. -> " + line, file);
                   throw std::runtime_error("too_much_regs");
                 }
               }
@@ -810,7 +810,7 @@ bool CompilerKit::EncoderPowerPC::WriteLine(std::string line, std::string file) 
                 ++found_some_count;
 
                 if (found_some_count > 1) {
-                  Detail::print_error("Too much registers. -> " + line, file);
+                  CompilerKit::Detail::print_error("Too much registers. -> " + line, file);
                   throw std::runtime_error("too_much_regs");
                 }
 
@@ -872,7 +872,7 @@ bool CompilerKit::EncoderPowerPC::WriteLine(std::string line, std::string file) 
 
           if (opcodeName == "mr") {
             if (register_count == 1) {
-              Detail::print_error("Too few registers. -> " + line, file);
+              CompilerKit::Detail::print_error("Too few registers. -> " + line, file);
               throw std::runtime_error("too_few_registers");
             }
           }
@@ -881,7 +881,7 @@ bool CompilerKit::EncoderPowerPC::WriteLine(std::string line, std::string file) 
           if (opcode_risc.ops->type != GREG) {
             // remember! register to register!
             if (found_some_count == 1) {
-              Detail::print_error(
+              CompilerKit::Detail::print_error(
                   "Unrecognized register found.\ntip: each AssemblerPower register "
                   "starts with 'r'.\nline: " +
                       line,
@@ -892,7 +892,7 @@ bool CompilerKit::EncoderPowerPC::WriteLine(std::string line, std::string file) 
           }
 
           if (found_some_count < 1 && name[0] != 'l' && name[0] != 's') {
-            Detail::print_error("invalid combination of opcode and registers.\nline: " + line,
+            CompilerKit::Detail::print_error("invalid combination of opcode and registers.\nline: " + line,
                                 file);
             throw std::runtime_error("invalid_comb_op_reg");
           }
