+++ src/CompilerKit/src/Compilers/CCompiler+Power64.cc
@@ -1,20 +1,16 @@
 /*
  *	========================================================
  *
- *	CCompilerARM64
+ *	CompilerPower64
  * 	Copyright (C) 2024-2025 Amlal El Mahrouss, Licensed under the Apache 2.0 license.
  *
  * 	========================================================
  */
 
-/// BUGS: 0
-/// TODO: none
-
 #include <CompilerKit/AST.h>
 #include <CompilerKit/UUID.h>
-#include <CompilerKit/Detail/Aarch64.h>
+#include <CompilerKit/Detail/Power64.h>
 #include <CompilerKit/Utilities/Compiler.h>
-
 #include <cstdio>
 #include <fstream>
 #include <iostream>
@@ -24,32 +20,20 @@
 #include <utility>
 #include <vector>
 
-/* C driver */
-/* This is part of the CompilerKit. */
-/* (c) Amlal El Mahrouss */
+#define kSourceExt ".c"
 
 /// @author Amlal El Mahrouss (amlal@nekernel.org)
-/// @file ARM64-cc.cc
-/// @brief ARM64 C Compiler.
-
-/// TODO: support structures, else if, else, . and  ->
-
-/////////////////////
-
-// ANSI ESCAPE CODES
-
-/////////////////////
-
-#define kCExtension ".c"
+/// @file cc.cc
+/// @brief POWER64 C Compiler.
 
 /////////////////////////////////////
 
-// INTERNAL STUFF OF THE C COMPILER
+/// INTERNAL STRUCTURES OF THE C COMPILER
 
 /////////////////////////////////////
 
 namespace Detail {
-// \brief Register map structure, used to keep track of each variable's registers.
+// \brief name to register struct.
 struct CompilerRegisterMap final {
   std::string fName;
   std::string fReg;
@@ -58,16 +42,16 @@ struct CompilerRegisterMap final {
 // \brief Map for C structs
 // \author amlal@nekernel.org
 struct CompilerStructMap final {
-  // 'my_foo'
+  /// 'struct::my_foo'
   std::string fName;
 
-  // if instance: stores a valid register.
+  /// if instance: stores a valid register.
   std::string fReg;
 
-  // offset count
+  /// offset count
   std::size_t fOffsetsCnt;
 
-  // offset array.
+  /// offset array.
   std::vector<std::pair<Int32, std::string>> fOffsets;
 };
 
@@ -110,7 +94,7 @@ static int kMachine = 0;
 /////////////////////////////////////////
 
 static size_t      kRegisterCnt     = kAsmRegisterLimit;
-static size_t      kStartUsable     = 8;
+static size_t      kStartUsable     = 2;
 static size_t      kUsableLimit     = 15;
 static size_t      kRegisterCounter = kStartUsable;
 static std::string kRegisterPrefix  = kAsmRegisterPrefix;
@@ -131,20 +115,20 @@ static bool                         kIfFound     = false;
 static size_t                       kBracesCount = 0UL;
 
 /* @brief C compiler backend for C */
-class CompilerFrontendARM64 final : public CompilerKit::CompilerFrontendInterface {
+class CompilerFrontendPower64 final : public CompilerKit::CompilerFrontendInterface {
  public:
-  explicit CompilerFrontendARM64()  = default;
-  ~CompilerFrontendARM64() override = default;
+  explicit CompilerFrontendPower64()  = default;
+  ~CompilerFrontendPower64() override = default;
 
-  NECTI_COPY_DEFAULT(CompilerFrontendARM64);
+  NECTI_COPY_DEFAULT(CompilerFrontendPower64);
 
   std::string                             Check(const char* text, const char* file);
   CompilerKit::SyntaxLeafList::SyntaxLeaf Compile(std::string text, std::string file) override;
 
-  const char* Language() override { return "64k C"; }
+  const char* Language() override { return "POWER C"; }
 };
 
-static CompilerFrontendARM64*            kCompilerFrontend = nullptr;
+static CompilerFrontendPower64*          kCompilerFrontend = nullptr;
 static std::vector<Detail::CompilerType> kCompilerVariables;
 static std::vector<std::string>          kCompilerFunctions;
 static std::vector<Detail::CompilerType> kCompilerTypes;
@@ -179,8 +163,10 @@ union double_cast final {
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
-CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendARM64::Compile(std::string text,
-                                                                       std::string file) {
+CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendPower64::Compile(std::string text_,
+                                                                         std::string file) {
+  std::string text = text_;
+
   bool typeFound = false;
   bool fnFound   = false;
 
@@ -194,7 +180,7 @@ CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendARM64::Compile(std::stri
 
   // start parsing
   for (size_t text_index = 0; text_index < text.size(); ++text_index) {
-    auto syntaxLeaf = CompilerKit::SyntaxLeafList::SyntaxLeaf();
+    auto syntax_leaf = CompilerKit::SyntaxLeafList::SyntaxLeaf();
 
     auto        gen = uuids::uuid_random_generator{generator};
     uuids::uuid out = gen();
@@ -225,9 +211,9 @@ CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendARM64::Compile(std::stri
               }
 
               if (text.find('(') != std::string::npos) {
-                syntaxLeaf.fUserValue = buf;
+                syntax_leaf.fUserValue = buf;
 
-                kState.fSyntaxTree->fLeafList.push_back(syntaxLeaf);
+                kState.fSyntaxTree->fLeafList.push_back(syntax_leaf);
               }
 
               typeFound = true;
@@ -250,7 +236,7 @@ CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendARM64::Compile(std::stri
       kInBraces = true;
       ++kBracesCount;
 
-      kState.fSyntaxTree->fLeafList.push_back(syntaxLeaf);
+      kState.fSyntaxTree->fLeafList.push_back(syntax_leaf);
     }
 
     // return keyword handler
@@ -292,17 +278,32 @@ CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendARM64::Compile(std::stri
             value += tmp;
           }
 
-          syntaxLeaf.fUserValue = "\tldw r19, ";
+          syntax_leaf.fUserValue = "\tmr r31, ";
 
           // make it pretty.
-          if (value.find('\t') != std::string::npos) value.erase(value.find('\t'), 1);
+          while (value.find('\t') != std::string::npos) value.erase(value.find('\t'), 1);
 
-          syntaxLeaf.fUserValue += value + "\n";
+          while (value.find(' ') != std::string::npos) value.erase(value.find(' '), 1);
+
+          while (value.find("extern_segment") != std::string::npos)
+            value.erase(value.find("extern_segment"), strlen("extern_segment"));
+
+          bool found = false;
+
+          for (auto& reg : kState.kStackFrame) {
+            if (value.find(reg.fName) != std::string::npos) {
+              found = true;
+              syntax_leaf.fUserValue += reg.fReg;
+              break;
+            }
+          }
+
+          if (!found) syntax_leaf.fUserValue += "r0";
         }
 
-        syntaxLeaf.fUserValue += "\tjlr";
+        syntax_leaf.fUserValue += "\n\tblr";
 
-        kState.fSyntaxTree->fLeafList.push_back(syntaxLeaf);
+        kState.fSyntaxTree->fLeafList.push_back(syntax_leaf);
 
         break;
       }
@@ -323,12 +324,14 @@ CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendARM64::Compile(std::stri
       kIfFunction = "__NECTI_IF_PROC_";
       kIfFunction += std::to_string(time_off._Raw);
 
-      syntaxLeaf.fUserValue = "\tlda r12, extern_segment ";
-      syntaxLeaf.fUserValue += kIfFunction +
-                               "\n\t#r12 = Code to jump on, r11 right cond, r10 left cond.\n\tbeq "
-                               "r10, r11, r12\ndword public_segment .code64 " +
-                               kIfFunction + "\n";
-      kState.fSyntaxTree->fLeafList.push_back(syntaxLeaf);
+      syntax_leaf.fUserValue =
+          "\tcmpw "
+          "r10, r11";
+
+      syntax_leaf.fUserValue += "\n\tbeq extern_segment " + kIfFunction +
+                                " \ndword public_segment .code64 " + kIfFunction + "\n";
+
+      kState.fSyntaxTree->fLeafList.push_back(syntax_leaf);
 
       kIfFound = true;
     }
@@ -345,7 +348,7 @@ CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendARM64::Compile(std::stri
       if (text.find("typedef ") != std::string::npos) continue;
 
       if (text[text_index] == '=' && kInStruct) {
-        Detail::print_error("assignement of value in struct " + text, file);
+        CompilerKit::Detail::print_error("assignement of value inside a struct " + text, file);
         continue;
       }
 
@@ -386,11 +389,11 @@ CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendARM64::Compile(std::stri
       if (text.find('=') != std::string::npos && kInBraces && !kIfFound) {
         if (text.find("*") != std::string::npos) {
           if (text.find("=") > text.find("*"))
-            substr += "\tlda ";
+            substr += "\tli ";
           else
-            substr += "\tldw ";
+            substr += "\tli ";
         } else {
-          substr += "\tldw ";
+          substr += "\tli ";
         }
       } else if (text.find('=') != std::string::npos && !kInBraces) {
         substr += "stw public_segment .data64 ";
@@ -471,22 +474,31 @@ CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendARM64::Compile(std::stri
           kCompilerVariables.cbegin(), kCompilerVariables.cend(),
           [&](Detail::CompilerType type) { return type.fName.find(substr) != std::string::npos; });
 
-      if (kRegisterCounter == 5 || kRegisterCounter == 6) ++kRegisterCounter;
+      kCompilerVariables.push_back({.fName = substr});
+
+      if (text[text_index] == ';') break;
 
       std::string reg = kAsmRegisterPrefix;
+
+      ++kRegisterCounter;
       reg += std::to_string(kRegisterCounter);
 
-      if (var_to_find == kCompilerVariables.cend()) {
-        ++kRegisterCounter;
+      auto newSubstr = substr.substr(substr.find(" "));
 
-        kState.kStackFrame.push_back({.fName = substr, .fReg = reg});
-        kCompilerVariables.push_back({.fName = substr});
+      std::string symbol;
+
+      for (size_t start = 0; start < newSubstr.size(); ++start) {
+        if (newSubstr[start] == ',') break;
+
+        if (newSubstr[start] == ' ') continue;
+
+        symbol += (newSubstr[start]);
       }
 
-      syntaxLeaf.fUserValue += substr;
-      kState.fSyntaxTree->fLeafList.push_back(syntaxLeaf);
+      kState.kStackFrame.push_back({.fName = symbol, .fReg = reg});
 
-      if (text[text_index] == '=') break;
+      syntax_leaf.fUserValue += "\n\tli " + reg + substr.substr(substr.find(','));
+      kState.fSyntaxTree->fLeafList.push_back(syntax_leaf);
     }
 
     // function handler.
@@ -533,7 +545,7 @@ CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendARM64::Compile(std::stri
           }
 
           args += args_buffer;
-          args += "\n\tlda r19, ";
+          args += "\n\tli r31, ";
         }
       }
 
@@ -553,22 +565,23 @@ CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendARM64::Compile(std::stri
       }
 
       if (kInBraces) {
-        syntaxLeaf.fUserValue = args;
-        syntaxLeaf.fUserValue += substr;
-        syntaxLeaf.fUserValue += "\n\tjrl\n";
+        syntax_leaf.fUserValue = args;
+        syntax_leaf.fUserValue += substr;
 
-        kState.fSyntaxTree->fLeafList.push_back(syntaxLeaf);
+        syntax_leaf.fUserValue += "\n\tblr\n";
+
+        kState.fSyntaxTree->fLeafList.push_back(syntax_leaf);
 
         fnFound = true;
       } else {
-        syntaxLeaf.fUserValue.clear();
+        syntax_leaf.fUserValue.clear();
 
-        syntaxLeaf.fUserValue += "public_segment .code64 ";
+        syntax_leaf.fUserValue += "public_segment .code64 ";
 
-        syntaxLeaf.fUserValue += substr;
-        syntaxLeaf.fUserValue += "\n";
+        syntax_leaf.fUserValue += substr;
+        syntax_leaf.fUserValue += "\n";
 
-        kState.fSyntaxTree->fLeafList.push_back(syntaxLeaf);
+        kState.fSyntaxTree->fLeafList.push_back(syntax_leaf);
 
         fnFound = true;
       }
@@ -583,10 +596,10 @@ CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendARM64::Compile(std::stri
         if (text[_text_i] == '\t' || text[_text_i] == ' ') text.erase(_text_i, 1);
       }
 
-      syntaxLeaf.fUserValue += "sub ";
-      syntaxLeaf.fUserValue += text;
+      syntax_leaf.fUserValue += "dec ";
+      syntax_leaf.fUserValue += text;
 
-      kState.fSyntaxTree->fLeafList.push_back(syntaxLeaf);
+      kState.fSyntaxTree->fLeafList.push_back(syntax_leaf);
       break;
     }
 
@@ -604,23 +617,23 @@ CompilerKit::SyntaxLeafList::SyntaxLeaf CompilerFrontendARM64::Compile(std::stri
 
       if (kInStruct) kInStruct = false;
 
-      kState.fSyntaxTree->fLeafList.push_back(syntaxLeaf);
+      kState.fSyntaxTree->fLeafList.push_back(syntax_leaf);
     }
 
-    syntaxLeaf.fUserValue.clear();
+    syntax_leaf.fUserValue.clear();
   }
 
-  auto syntaxLeaf       = CompilerKit::SyntaxLeafList::SyntaxLeaf();
-  syntaxLeaf.fUserValue = "\n";
-  kState.fSyntaxTree->fLeafList.push_back(syntaxLeaf);
+  auto syntax_leaf       = CompilerKit::SyntaxLeafList::SyntaxLeaf();
+  syntax_leaf.fUserValue = "\n";
+  kState.fSyntaxTree->fLeafList.push_back(syntax_leaf);
 
-  return syntaxLeaf;
+  return syntax_leaf;
 }
 
 static bool        kShouldHaveBraces = false;
 static std::string kFnName;
 
-std::string CompilerFrontendARM64::Check(const char* text, const char* file) {
+std::string CompilerFrontendPower64::Check(const char* text, const char* file) {
   std::string err_str;
   std::string ln = text;
 
@@ -1040,14 +1053,14 @@ skip_braces_check:
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
-class AssemblyCCInterface final CK_ASSEMBLY_INTERFACE {
+class AssemblyMountpointCLang final CK_ASSEMBLY_INTERFACE {
  public:
-  explicit AssemblyCCInterface()  = default;
-  ~AssemblyCCInterface() override = default;
+  explicit AssemblyMountpointCLang()  = default;
+  ~AssemblyMountpointCLang() override = default;
 
-  NECTI_COPY_DEFAULT(AssemblyCCInterface);
+  NECTI_COPY_DEFAULT(AssemblyMountpointCLang);
 
-  UInt32 Arch() noexcept override { return CompilerKit::AssemblyFactory::kArchAARCH64; }
+  UInt32 Arch() noexcept override { return CompilerKit::AssemblyFactory::kArchPowerPC; }
 
   Int32 CompileToFormat(std::string src, Int32 arch) override {
     if (kCompilerFrontend == nullptr) return 1;
@@ -1074,7 +1087,7 @@ class AssemblyCCInterface final CK_ASSEMBLY_INTERFACE {
     auto fmt = CompilerKit::current_date();
 
     (*kState.fOutputAssembly) << "# Path: " << src_file << "\n";
-    (*kState.fOutputAssembly) << "# Language: ARM64 Assembly (Generated from ANSI C)\n";
+    (*kState.fOutputAssembly) << "# Language: POWER Assembly (Generated from C)\n";
     (*kState.fOutputAssembly) << "# Date: " << fmt << "\n\n";
 
     CompilerKit::SyntaxLeafList syntax;
@@ -1088,13 +1101,13 @@ class AssemblyCCInterface final CK_ASSEMBLY_INTERFACE {
       if (auto err = kCompilerFrontend->Check(line_src.c_str(), src.data()); err.empty()) {
         kCompilerFrontend->Compile(line_src, src.data());
       } else {
-        Detail::print_error(err, src.data());
+        CompilerKit::Detail::print_error(err, src.data());
       }
     }
 
     if (kAcceptableErrors > 0) return 1;
 
-    std::vector<std::string> keywords = {"ldw", "stw", "lda", "sta", "add", "sub", "mv"};
+    std::vector<std::string> keywords = {"ld", "stw", "add", "sub", "or"};
 
     ///
     /// Replace, optimize, fix assembly output.
@@ -1137,10 +1150,9 @@ class AssemblyCCInterface final CK_ASSEMBLY_INTERFACE {
             }
 
             if (CompilerKit::find_word(leaf.fUserValue, needle)) {
-              if (leaf.fUserValue.find("extern_segment " + needle) != std::string::npos) {
-                std::string range = "extern_segment " + needle;
-                leaf.fUserValue.replace(leaf.fUserValue.find("extern_segment " + needle),
-                                        range.size(), needle);
+              if (leaf.fUserValue.find("extern_segment ") != std::string::npos) {
+                std::string range = "extern_segment ";
+                leaf.fUserValue.replace(leaf.fUserValue.find(range), range.size(), "");
               }
 
               if (leaf.fUserValue.find("ldw r6") != std::string::npos) {
@@ -1148,7 +1160,7 @@ class AssemblyCCInterface final CK_ASSEMBLY_INTERFACE {
                     std::count(leaf.fUserValue.begin(), leaf.fUserValue.end(), ',');
 
                 if (countComma == 1) {
-                  leaf.fUserValue.replace(leaf.fUserValue.find("ldw"), strlen("ldw"), "mv");
+                  leaf.fUserValue.replace(leaf.fUserValue.find("ldw"), strlen("ldw"), "mr");
                 }
               }
 
@@ -1158,8 +1170,8 @@ class AssemblyCCInterface final CK_ASSEMBLY_INTERFACE {
             }
           }
 
-          if (cnt > 1 && keyword != "mv" && keyword != "add" && keyword != "sub") {
-            leaf.fUserValue.replace(leaf.fUserValue.find(keyword), keyword.size(), "mv");
+          if (cnt > 1 && keyword != "mr" && keyword != "add" && keyword != "dec") {
+            leaf.fUserValue.replace(leaf.fUserValue.find(keyword), keyword.size(), "mr");
           }
         }
       }
@@ -1178,8 +1190,8 @@ class AssemblyCCInterface final CK_ASSEMBLY_INTERFACE {
   }
 };
 
-NECTI_MODULE(CompilerCLangARM64) {
-  ::signal(SIGSEGV, Detail::drvi_crash_handler);
+NECTI_MODULE(CompilerCLangPowerPC) {
+  CompilerKit::install_signal(SIGSEGV, CompilerKit::Detail::drvi_crash_handler);
 
   kCompilerTypes.push_back({.fName = "void", .fValue = "void"});
   kCompilerTypes.push_back({.fName = "char", .fValue = "byte"});
@@ -1190,9 +1202,9 @@ NECTI_MODULE(CompilerCLangARM64) {
 
   bool skip = false;
 
-  kFactory.Mount(new AssemblyCCInterface());
-  kMachine          = CompilerKit::AssemblyFactory::kArchAARCH64;
-  kCompilerFrontend = new CompilerFrontendARM64();
+  kFactory.Mount(new AssemblyMountpointCLang());
+  kMachine          = CompilerKit::AssemblyFactory::kArchPowerPC;
+  kCompilerFrontend = new CompilerFrontendPower64();
 
   for (auto index = 1UL; index < argc; ++index) {
     if (skip) {
@@ -1201,27 +1213,29 @@ NECTI_MODULE(CompilerCLangARM64) {
     }
 
     if (argv[index][0] == '-') {
-      if (strcmp(argv[index], "--v") == 0 || strcmp(argv[index], "--version") == 0) {
+      if (strcmp(argv[index], "-v") == 0 || strcmp(argv[index], "-version") == 0) {
+        // AMLALE Compiler Version.
         return EXIT_SUCCESS;
       }
 
-      if (strcmp(argv[index], "--verbose") == 0) {
+      if (strcmp(argv[index], "-verbose") == 0) {
         kState.fVerbose = true;
 
         continue;
       }
 
-      if (strcmp(argv[index], "--h") == 0 || strcmp(argv[index], "--help") == 0) {
+      if (strcmp(argv[index], "-h") == 0 || strcmp(argv[index], "-help") == 0) {
+        // AMLALE Compiler Help.
         return EXIT_SUCCESS;
       }
 
-      if (strcmp(argv[index], "--dialect") == 0) {
+      if (strcmp(argv[index], "-dialect") == 0) {
         if (kCompilerFrontend) std::cout << kCompilerFrontend->Language() << "\n";
 
         return EXIT_SUCCESS;
       }
 
-      if (strcmp(argv[index], "--fmax-exceptions") == 0) {
+      if (strcmp(argv[index], "-fmax-exceptions") == 0) {
         try {
           kErrorLimit = std::strtol(argv[index + 1], nullptr, 10);
         }
@@ -1238,7 +1252,7 @@ NECTI_MODULE(CompilerCLangARM64) {
       std::string err = "Unknown command: ";
       err += argv[index];
 
-      Detail::print_error(err, "cc");
+      CompilerKit::Detail::print_error(err, "cc");
 
       continue;
     }
@@ -1247,9 +1261,9 @@ NECTI_MODULE(CompilerCLangARM64) {
 
     std::string srcFile = argv[index];
 
-    if (strstr(argv[index], kCExtension) == nullptr) {
+    if (strstr(argv[index], kSourceExt) == nullptr) {
       if (kState.fVerbose) {
-        Detail::print_error(srcFile + " is not a valid C source.\n", "cc");
+        CompilerKit::Detail::print_error(srcFile + " is not a valid C source.\n", "cc");
       }
 
       return 1;
